requirements.txt
requests==2.26.0
dnspython==2.2.0